Revert most of texture "fixes".
Need to completely rethink how to do this.
Turns out it's actually legal in legacy GL (and GLES2) to call glBindTexture with a texture
id that's not generated by glGenTextures. However, this is very legacy
behavior, and as legacy as GL can get.

Still breaks with FBO textures however, so keep the "mapping" scheme for that
until a better solution is in place.

The major cause of texture id balooning is that the glide wrapper seems
to cycle through all available memory in a huge ring. This causes every new
texture upload to create a new texture. Can easily cause ~3000+++
textures to be uploaded which seems to cause some serious performance
issues, especially in sprite heavy games.

PBOs with a fixed cache of texture IDs might be a better solution
overall. It would map much more cleanly to this model at least ...