Joystick range takes the dead-zone into account
This change scales the joystick range to include the deadzone. This allows the full -0x8000 to 0x8000 range to be read, with near-zero values right outside the deadzone (instead of jumping to 0x1000 when the deadzone is passed). Walking slowly and sneaking (like in Banjo Tooie, where you need to get past Sssslumber and the cave man), is now possible.