corrected use of nonstandard `inline` extension for portability
`inline` is not a valid C keyboard in standard C compilers except perhaps those that have a mostly featured C99 extension.  For example, MSVC uses `__inline`, not `inline` (while in MSVC++ the latter is acceptable).  Since in-lining functions is not really a part of C programming itself (but more a compiler hint for how to allocate the algorithm) it is best left to a macro that different compilers may have to resolve on their own.  In the worst case, there is no inline feature at all to define the macro to, but the compiler can still handle inlining by itself without a hint.