vulkan_symbol_wrapper.cpp - this gets me ingame with Super Mario 64 on Ivy Bridge Vulkan - it shows the Mario 64 title screen, it zooms out, you hear the mario sound, then a black screen.
jekstrand notes on this file -
Heads up: Most of those symbols aren't instance symbols.
You should really be using GetDeviceProcAddress for most of them.

Basically everything other than the things for creating/destroying
isntances, physical devices,a nd devices.
Those should be GetInstanceProcAddress and the others should be
GetDeviceProCAddress
I don't think it matters in our [Intel Mesa] driver.

Bah, the Vulkan spec is extremely unclear about which ones are instance
and which aren't because it assumes you're going through the loader.
The short version is that you use GetInstanceProcAddress only for those
things which you need to use without a device and GetDeviceProcAddress
for the rest
As long as you're going throgh the loader, GetInstanceProcAddresswill
work for everything but the pinters you get may not be as efficient

Commenting on the python generation script -
update entry_is_device