Added 4:3 resolution for 4K & 8K screens
There are probably users with 4K screens (3840x2160) using libretro/RetroArch now, and in a near future users with even 8K screens (7680x4320).

I used some online aspect ratio calculators to find the 4:3 variant of those 16:9 resolutions (as the 4:3 aspect ratio is most common in N64 games). The 4:3 variant of 4K is 2880x2160, and the 4:3 variant of 8K is 5760x4320.