fixed a typo from the update to VCH's complements conversion
While patching the VCH op-code on the RSP to account for the corner case scenario of (VT == -32768), a typo was made that caused the updated VT temporary to be decremented by 1 in the presence of the corner case scenario, when the intention was to only continue to decrement by 1 (as the original algorithm did) in the absence of said scenario, not the presence.  A simple bit-wise NOT corrects the inversion.

The bug was a simple off-by-one inaccuracy with some of the vector clipping results.  No visible bugs have been noticed and attributed from within games to this regression; the only reason I noticed it is because Zapeth ran the op-code under test ROMs on the real N64 while LegendOfDragoon pin-pointed the offending commit.